feat: create new component for badge service status
