+++ src/components/Blocks/Listing/SimpleCard/Card/SimpleCardDefault.jsx
@@ -9,7 +9,6 @@ import {
   CardTitle,
   CardText,
   CardReadMore,
-  Badge,
 } from 'design-react-kit';
 import { injectLazyLibs } from '@plone/volto/helpers/Loadable/Loadable';
 import UniversalLink from '@plone/volto/components/manage/UniversalLink/UniversalLink';
@@ -22,6 +21,7 @@ import {
 } from 'io-sanita-theme/helpers';
 import { CardCategoryBottom } from 'io-sanita-theme/components';
 import { ListingText } from 'io-sanita-theme/components/Blocks';
+import { BadgeStatusServizio } from 'io-sanita-theme/components/View/Servizio';
 
 import { CardCategoryTop } from 'io-sanita-theme/components';
 import 'io-sanita-theme/components/Cards/CardSimple/cardSimple.scss';
@@ -35,10 +35,6 @@ const messages = defineMessages({
   update_date: {
     id: 'update_date',
     defaultMessage: 'Data di aggiornamento',
-  },
-  servizioOnline: {
-    id: 'servizio_online_chip',
-    defaultMessage: 'Servizio online',
   },
 });
 
@@ -129,9 +125,7 @@ const SimpleCardDefault = (props) => {
           {/* Chip servizio attivo */}
           {isServizioOnline && (
             <div className="mb-3">
-              <Badge color="info">
-                {intl.formatMessage(messages.servizioOnline)}
-              </Badge>
+              <BadgeStatusServizio servizio_attivo={true} />
             </div>
           )}
           {listingText && (
