+++ src/components/View/commons/PageHeader/PageHeaderStatoServizio.jsx
@@ -1,18 +1,10 @@
 import { defineMessages, useIntl } from 'react-intl';
-import { Badge } from 'design-react-kit';
+import { BadgeStatusServizio } from 'io-sanita-theme/components/View/Servizio';
 
 const messages = defineMessages({
   service_available: {
     id: 'Service available',
     defaultMessage: 'Servizio disponibile e prenotabile',
-  },
-  service_on: {
-    id: 'service_on',
-    defaultMessage: 'Servizio attivo',
-  },
-  service_off: {
-    id: 'service_off',
-    defaultMessage: 'Servizio non attivo',
   },
 });
 
@@ -21,11 +13,7 @@ const PageHeaderServizio = ({ content }) => {
 
   return content['@type'] === 'Servizio' ? (
     <div className="mb-3">
-      <Badge color={content.servizio_attivo ? 'info' : 'danger'}>
-        {content.servizio_attivo
-          ? intl.formatMessage(messages.service_on)
-          : intl.formatMessage(messages.service_off)}
-      </Badge>
+      <BadgeStatusServizio servizio_attivo={content.servizio_attivo} />
       {content.servizio_attivo && (
         <span className="px-2">
           {intl.formatMessage(messages.service_available)}
